requirements.txt
# Core dependencies
python-dotenv==1.1.0
requests==2.32.3
rich==14.0.0
together==1.5.8

# Web dependencies
flask==3.0.3
flask-cors==4.0.1

# Dependencies (automatically installed)
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
annotated-types==0.7.0
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
eval-type-backport==0.2.2
filelock==3.18.0
frozenlist==1.6.0
idna==3.10
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.4
numpy==2.2.6
pillow==11.2.1
propcache==0.3.1
pyarrow==20.0.0
pydantic==2.11.5
pydantic-core==2.33.2
pygments==2.19.1
shellingham==1.5.4
tabulate==0.9.0
tqdm==4.67.1
typer==0.15.4
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
yarl==1.20.0

# Optional packages (uncomment if needed)
# prompt-toolkit==3.0.39  # For interactive prompts
# tiktoken==0.5.1  # For token counting

